init
in this commit the program is able to work with the first PDF we received.

some work need to be done to make it presentable at the end.

some work will be needed to transfer from a text file to the json so we can transfer our work

last task to do before next week will be to make a way that all the answer look pretty in a word document for when its over.

Maybe make an interaction menue with more options ?

Luxury_Karma